3lo.test/requirements.txt
# Requirements for Google Calendar OAuth2 3LO test
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
